requirements.txt
streamlit
pandas
matplotlib
openpyxl
fpdf2
openai
google-generativeai